Fix GeoObject is now where it should be, at domain, its new home. Fix Validator now properly handles 301 and 429 codes.
