+++ common/src/main/java/urlshortener/common/services/UrlValidatorAndCheckerImpl.java
@@ -3,10 +3,7 @@ package urlshortener.common.services;
 import org.apache.commons.validator.routines.UrlValidator;
 
 import java.io.IOException;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.ProtocolException;
-import java.net.URL;
+import java.net.*;
 
 public class UrlValidatorAndCheckerImpl implements UrlValidatorAndChecker {
 
@@ -23,28 +20,39 @@ public class UrlValidatorAndCheckerImpl implements UrlValidatorAndChecker {
 
     @Override
     public  boolean isAlive(String url){
-        int code = -1;
-        try {
-        URL urlConnection = new URL(url);
-        HttpURLConnection connection = (HttpURLConnection)urlConnection.openConnection();
-        connection.setRequestMethod("GET");
-        connection.connect();
-        code = connection.getResponseCode();
-        } catch (ProtocolException e) {
-            e.printStackTrace();
-        } catch (MalformedURLException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            if (code == 200){
-                return true;
-            }else{
-                return false;
-            }
-        }
-            if (code == 200){
+        int code = getCode(url);
+        System.out.println(code);
+        if (code == 200){
             return true;
-        }else{
+        }
+        else{
+            System.out.println("link muerto");
             return false;
         }
+    }
+
+    private int getCode(String url) {
+        int code = -1;
+        try {
+            URL urlConnection = new URL(url);
+            HttpURLConnection connection = (HttpURLConnection) urlConnection.openConnection();
+            connection.setRequestMethod("GET");
+            connection.connect();
+            code = connection.getResponseCode();
+            if (code == 301) {
+                System.out.println("recibido 301, redireccionando");
+                code = getCode(connection.getHeaderField("location"));
+            }
+            if (code == 429) {
+                System.out.println("recibido 429, reintentando");
+                code = getCode(url);
+            }
+        } catch (ProtocolException | MalformedURLException e) {
+            System.out.println("link malo");
+            e.printStackTrace();
+        } catch (IOException e) {
+            return code;
+        }
+        return code;
     }
 }
